backend/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.10.1
PyJWT==2.8.0
certifi==2025.10.5
unidecode==1.4.0

# Authentication
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-multipart==0.0.6

# MongoDB
pymongo==4.6.1
motor==3.3.2

# Environment
python-dotenv==1.0.0

# Email (opcjonalnie - możesz tymczasowo zakomentować)
fastapi-mail==1.5.0